requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
jwt==1.3.1
Mako==1.3.5
MarkupSafe==3.0.1
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-jose==3.3.0
python-multipart==0.0.12
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.1
